only execute before_deploy travis script once

+++ scripts/travis_build.sh
@@ -1,17 +1,21 @@
 # Abort on any error
 set -e
 
-# Build Frontend Code
-cd frontend
-npm run build
+# Only run before first deploy
+if ! [ "$BEFORE_DEPLOY_RUN" ]; then
+  export BEFORE_DEPLOY_RUN=1;
+  # Build Frontend Code
+  cd frontend
+  npm run build
 
-# Return to root directory
-cd ..
+  # Return to root directory
+  cd ..
 
-# Remove node_modules from deployment
-rm -rf frontend/node_modules
+  # Remove node_modules from deployment
+  rm -rf frontend/node_modules
 
-# Package files together to upload to S3
-zip -r latest *
-mkdir -p dpl_cd_upload
-mv latest.zip dpl_cd_upload/latest.zip
+  # Package files together to upload to S3
+  zip -r latest *
+  mkdir -p dpl_cd_upload
+  mv latest.zip dpl_cd_upload/latest.zip
+fi
